requirements.txt
Flask==2.1.1
torch==1.11.0
numpy==1.22.3
Pillow==9.1.0
torchvision==0.12.0
opencv-python==4.5.5.64
rembg==2.0.30
